third_party/blink/web_tests/hittesting/text-overflow-inline-image-expected.txt

CONSOLE MESSAGE: A soft navigation has been detected: text-overflow-inline-image.html#
this is a link
clicked
